Added the folder- Took everything from previous project. No change afaik. Don't know why it worked here tbh.
So, the previous github repository has afaik the same code, don't remember if i changed anything here- maybe i did? But the code seems to be working here. Maybe something went wrong with AS project or something with the emulator. 
The reason i won't have commits here, is because i already have the finished code. Good thing I was updating the project in another AS project. Otherwise I'd be doomed.